works but prediction bug causes SLIGHT mispredict, likely needs an RT pull request.
