+++ Content.Client/_Starlight/Contaniers/ThrowInsertContainerSystem.cs
@@ -1,0 +1,5 @@
+using Content.Shared._Starlight.Containers;
+
+namespace Content.Client._Starlight.Contaniers;
+
+public sealed class ThrowInsertContainerSystem : SharedThrowInsertContainerSystem;
