+++ Content.Server/Containers/ThrowInsertContainerSystem.cs
@@ -6,50 +6,8 @@ using Content.Shared.Throwing;
 using Robust.Shared.Audio.Systems;
 using Robust.Shared.Containers;
 using Robust.Shared.Random;
+using Content.Shared._Starlight.Containers;
 
 namespace Content.Server.Containers;
 
-public sealed partial class ThrowInsertContainerSystem : EntitySystem
-{
-    [Dependency] private IAdminLogManager _adminLogger = default!;
-    [Dependency] private SharedAudioSystem _audio = default!;
-    [Dependency] private SharedContainerSystem _containerSystem = default!;
-    [Dependency] private SharedPopupSystem _popup = default!;
-    [Dependency] private IRobustRandom _random = default!;
-
-    public override void Initialize()
-    {
-        base.Initialize();
-
-        SubscribeLocalEvent<ThrowInsertContainerComponent, ThrowHitByEvent>(OnThrowCollide);
-    }
-
-    private void OnThrowCollide(Entity<ThrowInsertContainerComponent> ent, ref ThrowHitByEvent args)
-    {
-        var container = _containerSystem.GetContainer(ent, ent.Comp.ContainerId);
-
-        if (!_containerSystem.CanInsert(args.Thrown, container))
-            return;
-
-        var beforeThrowArgs = new BeforeThrowInsertEvent(args.Thrown);
-        RaiseLocalEvent(ent, ref beforeThrowArgs);
-
-        if (beforeThrowArgs.Cancelled)
-            return;
-
-        if (!_random.Prob(ent.Comp.Probability))
-        {
-            _audio.PlayPvs(ent.Comp.MissSound, ent);
-            _popup.PopupEntity(Loc.GetString(ent.Comp.MissLocString), ent);
-            return;
-        }
-
-        if (!_containerSystem.Insert(args.Thrown, container))
-            throw new InvalidOperationException("Container insertion failed but CanInsert returned true");
-
-        _audio.PlayPvs(ent.Comp.InsertSound, ent);
-
-        if (args.Component.Thrower != null)
-            _adminLogger.Add(LogType.Landed, LogImpact.Low, $"{ToPrettyString(args.Thrown)} thrown by {ToPrettyString(args.Component.Thrower.Value):player} landed in {ToPrettyString(ent)}");
-    }
-}
+public sealed class ThrowInsertContainerSystem : SharedThrowInsertContainerSystem; // Starlight edit: Moved to shared.
